Issue-6 fixed recursion working now we can parse the whole config
There are a couple issues still, we're not handling non-string scalar
types (numbers) and we're not passing the name of the key to the
callback function that caused us to go into a reference.

This second issue is a little trickier to handle, I think I will
add a member to the struct keydata that holds the outter key value in
the hash that represents the current reference we are recursing into.

       struct keydata {
         enum { ... } type;
         int index;
         const char * key;
         const char * parent_key;  /* The key in our parent hash (if there is
       one) */
       };

       ie: if there is a hash like so:
       my %hash = (
          elements => {
              'user' => 'blah',
              'pass' => 12345,
          }
       );

Then if we are inspecting the 'user' key in the 'elements' hashreference
'key' will be 'user' and 'parent_key' will be 'elements'