Issue-8 initial commit
Adding user module to keep track of users emails/identify them for
access levels later on and to be able to send emails to them and what
not. I have documented the module so you can see how to use
the user library which is an abstraction over the TokyoCabinet::TDB
database format.