Partial changes toward fixing the 'reload' problem
Reloading oesn't work because the child handling doesn't seem to work
properly as far as children getting reaped and then respawned correctly
and keeping track of the sockets involved when that happens.

How it should work:
pids are stored in a hash table. like so:
pid => {
  sock => fd    # the socket we're using to communicate with this pid on (created by socketpair())
  ...
}

When the parent catches a SIGCHLD (which may have been compressed
and thus we should probably keep waiting asynchronously until
we get an error)-- it will get the pid of that child
(or children) and delete their entries in the hash table. It will
then spawn off as many new children as necessary to get
back to the max number of worker children. In doing so it must
also ensure to reopen the socketpair IPC communication between
the parent and child correctly.