Linkbot updates make it slightly more cachable (although no cache yet)
This fixes the exemption checking so its now an O(1) lookup instead
of O(N) where N whas the number of the exemptions to check. This also
improves the creation of the exemption and threat_checking functions
to be factory design.

Some future improvements that I plan to do for this module:
* Create a sub-module system to act differently depending on the
site. For example a special module for 'youtoube.com' to show video
and description, or even a direct download link (Via youtube-dl)
or show special informatoin if its a news site
* Create a shared cache between processes that stores the threat
checking information for a given domain so we don't keep looking it up
over and over (this happens currently..)