Add BROADCAST command to children which basically allows one child to re-broadcast a received line to all children
This is a useful way to send messages to all children for certain messages. One example (and the use case we created this for)
is if a command is received by a child that requires the child to register an event handler we can broadcast that command
which will then make all children register that same event handler with the same parameters, allowing us to register and
remove event handlers on the fly at runtime, instead of just at startup.

NOTE: The child broadcasting the message will ALSO receive the message, for the use case above that won't matter
since it will just result in the same handler being registered again, overwriting the last one, there may be some use
cases though where you want to catch that duplicate and ignore it in the child doing the broadcasting.